Proyecto_Integador/requirements.txt
customtkinter
Pillow
matplotlib
json
mysql-connector-python